refactorized api
